refactor: public/data -> src/mocks
