+++ pages/index.jsx
@@ -1,5 +1,5 @@
 export { default } from '@/home'
-import courses from '@/public/data/courses.json'
+import courses from '@/mocks/courses.json'
 
 export async function getStaticProps() {
   return {
