--- public/data/users.json
@@ -1,242 +1,0 @@
-[
-  {
-    "id": "60ac623442cadab122c0c6fa",
-    "avatar": "https://picsum.photos/64.webp?random=6",
-    "username": "DaphneClay",
-    "role": "Lecturer",
-    "password": "dolor proident",
-    "email": "daphneclay@zosis.net"
-  },
-  {
-    "id": "60ac623448b8b6f63e5edc95",
-    "avatar": "https://picsum.photos/64.webp?random=21",
-    "username": "ReyesBentley",
-    "role": "Student",
-    "password": "enim quis",
-    "email": "reyesbentley@entropix.name"
-  },
-  {
-    "id": "60ac6234fd8a709a5abeb06e",
-    "avatar": "https://picsum.photos/64.webp?random=4",
-    "username": "StokesJackson",
-    "role": "Lecturer",
-    "password": "tempor nostrud",
-    "email": "stokesjackson@quonata.org"
-  },
-  {
-    "id": "60ac6234db7d9242a68afd3f",
-    "avatar": "https://picsum.photos/64.webp?random=12",
-    "username": "MccartySanford",
-    "role": "Admin",
-    "password": "laboris nisi",
-    "email": "mccartysanford@aquazure.me"
-  },
-  {
-    "id": "60ac6234e2d5d68ae6154a97",
-    "avatar": "https://picsum.photos/64.webp?random=25",
-    "username": "IngramFarmer",
-    "role": "Lecturer",
-    "password": "dolor fugiat",
-    "email": "ingramfarmer@rodemco.tv"
-  },
-  {
-    "id": "60ac6234210538988f8448ad",
-    "avatar": "https://picsum.photos/64.webp?random=16",
-    "username": "SteeleMiller",
-    "role": "Admin",
-    "password": "in consequat",
-    "email": "steelemiller@anocha.com"
-  },
-  {
-    "id": "60ac623465d541e9ace88e0b",
-    "avatar": "https://picsum.photos/64.webp?random=17",
-    "username": "HintonNewton",
-    "role": "Admin",
-    "password": "ullamco ex",
-    "email": "hintonnewton@zorromop.biz"
-  },
-  {
-    "id": "60ac6234f492adcaef1bb322",
-    "avatar": "https://picsum.photos/64.webp?random=28",
-    "username": "VanessaBishop",
-    "role": "Student",
-    "password": "adipisicing reprehenderit",
-    "email": "vanessabishop@acumentor.io"
-  },
-  {
-    "id": "60ac6234965b4c6ccfc99177",
-    "avatar": "https://picsum.photos/64.webp?random=17",
-    "username": "CoxHowe",
-    "role": "Admin",
-    "password": "irure aute",
-    "email": "coxhowe@geostele.ca"
-  },
-  {
-    "id": "60ac6234b9687e0bcc554276",
-    "avatar": "https://picsum.photos/64.webp?random=5",
-    "username": "EvangelineMathis",
-    "role": "Lecturer",
-    "password": "sunt occaecat",
-    "email": "evangelinemathis@zillactic.us"
-  },
-  {
-    "id": "60ac6234a01ccb82af933695",
-    "avatar": "https://picsum.photos/64.webp?random=17",
-    "username": "ParksDaugherty",
-    "role": "Lecturer",
-    "password": "ad excepteur",
-    "email": "parksdaugherty@qualitex.biz"
-  },
-  {
-    "id": "60ac6234ada853939544c83f",
-    "avatar": "https://picsum.photos/64.webp?random=2",
-    "username": "TerrieTyson",
-    "role": "Student",
-    "password": "ullamco aliqua",
-    "email": "terrietyson@kangle.info"
-  },
-  {
-    "id": "60ac62340d95c29c59b761c6",
-    "avatar": "https://picsum.photos/64.webp?random=13",
-    "username": "HoustonJennings",
-    "role": "Admin",
-    "password": "tempor anim",
-    "email": "houstonjennings@digirang.net"
-  },
-  {
-    "id": "60ac62340d515809220b2065",
-    "avatar": "https://picsum.photos/64.webp?random=15",
-    "username": "LloydPreston",
-    "role": "Lecturer",
-    "password": "aute ut",
-    "email": "lloydpreston@phormula.name"
-  },
-  {
-    "id": "60ac6234ca0167e11f0732b2",
-    "avatar": "https://picsum.photos/64.webp?random=8",
-    "username": "HolderHart",
-    "role": "Admin",
-    "password": "in excepteur",
-    "email": "holderhart@magneato.org"
-  },
-  {
-    "id": "60ac6234057d28adbf878211",
-    "avatar": "https://picsum.photos/64.webp?random=22",
-    "username": "HuntMccarty",
-    "role": "Student",
-    "password": "qui pariatur",
-    "email": "huntmccarty@pyramia.me"
-  },
-  {
-    "id": "60ac6234a6d6cd43d822d7a3",
-    "avatar": "https://picsum.photos/64.webp?random=15",
-    "username": "CorineBurnett",
-    "role": "Student",
-    "password": "sit deserunt",
-    "email": "corineburnett@gallaxia.tv"
-  },
-  {
-    "id": "60ac6234a221deb51ec432c5",
-    "avatar": "https://picsum.photos/64.webp?random=25",
-    "username": "BauerStout",
-    "role": "Admin",
-    "password": "consequat est",
-    "email": "bauerstout@exostream.com"
-  },
-  {
-    "id": "60ac623424538c410ef2d6d5",
-    "avatar": "https://picsum.photos/64.webp?random=10",
-    "username": "DoyleVaughan",
-    "role": "Admin",
-    "password": "eu dolore",
-    "email": "doylevaughan@portico.biz"
-  },
-  {
-    "id": "60ac6234996674257199b194",
-    "avatar": "https://picsum.photos/64.webp?random=1",
-    "username": "ZamoraFerrell",
-    "role": "Lecturer",
-    "password": "reprehenderit fugiat",
-    "email": "zamoraferrell@pathways.io"
-  },
-  {
-    "id": "60ac6234a81e03a55b3dc704",
-    "avatar": "https://picsum.photos/64.webp?random=14",
-    "username": "AngelinaWiley",
-    "role": "Admin",
-    "password": "laboris magna",
-    "email": "angelinawiley@uncorp.ca"
-  },
-  {
-    "id": "60ac62340eed04e7b3b73d77",
-    "avatar": "https://picsum.photos/64.webp?random=9",
-    "username": "DellaHuber",
-    "role": "Admin",
-    "password": "dolore exercitation",
-    "email": "dellahuber@balooba.us"
-  },
-  {
-    "id": "60ac6234c3dae2af64b9c01e",
-    "avatar": "https://picsum.photos/64.webp?random=24",
-    "username": "JeannineMcfarland",
-    "role": "Admin",
-    "password": "occaecat ipsum",
-    "email": "jeanninemcfarland@bitrex.biz"
-  },
-  {
-    "id": "60ac62344eacefa00fe12481",
-    "avatar": "https://picsum.photos/64.webp?random=1",
-    "username": "GoffTownsend",
-    "role": "Lecturer",
-    "password": "ullamco duis",
-    "email": "gofftownsend@mantrix.info"
-  },
-  {
-    "id": "60ac62349c0cb4f54649e793",
-    "avatar": "https://picsum.photos/64.webp?random=15",
-    "username": "MyraOneil",
-    "role": "Lecturer",
-    "password": "consequat officia",
-    "email": "myraoneil@geeketron.net"
-  },
-  {
-    "id": "60ac6234b482fae9c71cc3c9",
-    "avatar": "https://picsum.photos/64.webp?random=17",
-    "username": "CharmaineNoble",
-    "role": "Admin",
-    "password": "sint eiusmod",
-    "email": "charmainenoble@boilcat.name"
-  },
-  {
-    "id": "60ac6234a1b6fdfb9df5c6c2",
-    "avatar": "https://picsum.photos/64.webp?random=19",
-    "username": "BenderJensen",
-    "role": "Lecturer",
-    "password": "commodo ipsum",
-    "email": "benderjensen@marqet.org"
-  },
-  {
-    "id": "60ac6234ef2b3f4ee099b0ad",
-    "avatar": "https://picsum.photos/64.webp?random=6",
-    "username": "ClaireDominguez",
-    "role": "Student",
-    "password": "esse labore",
-    "email": "clairedominguez@pushcart.me"
-  },
-  {
-    "id": "60ac62349b7ae615c6f5a9b3",
-    "avatar": "https://picsum.photos/64.webp?random=14",
-    "username": "DebraMelton",
-    "role": "Lecturer",
-    "password": "eiusmod labore",
-    "email": "debramelton@brainclip.tv"
-  },
-  {
-    "id": "60ac6234c690dba80c6949ab",
-    "avatar": "https://picsum.photos/64.webp?random=8",
-    "username": "FannieFields",
-    "role": "Lecturer",
-    "password": "est velit",
-    "email": "fanniefields@zilla.com"
-  }
-]
